+++ mercator-core/src/main/java/org/lendingclub/mercator/core/Scanner.java
@@ -4,4 +4,5 @@ public interface Scanner {
 
 	public void scan();
 	public Projector getProjector();
+	public SchemaManager getSchemaManager();
 }